build(Makefile): update build strategy
The old container method of building this site no longer works as the
container that we were using hasn't been updated in a bit and doesn't
seem to work with newer gems.

This introduces a dockerfile that we own and allows us to be a bit more
flexible in the way that we build the site. Additionally, since the
container will be built on the host, it means that people with different
machine architectures should be able to build the site as well.